# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Academy Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register student or faculty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [auth]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List all admins
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Admin list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Create a new admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Created admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}:
    get:
      operationId: getAdmin
      tags: [admins]
      summary: Get admin by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateAdmin
      tags: [admins]
      summary: Update admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdate"
      responses:
        "200":
          description: Updated admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Delete admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admins/{id}/status:
    patch:
      operationId: updateAdminStatus
      tags: [admins]
      summary: Block / unblock / activate / deactivate admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admins/{id}/reset-password:
    post:
      operationId: resetAdminPassword
      tags: [admins]
      summary: Reset admin password
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset

  # ── Courses ───────────────────────────────────────────────────────────────
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List batches
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Batch list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [batches]
      summary: Get batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    patch:
      operationId: updateBatch
      tags: [batches]
      summary: Update batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpdate"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    delete:
      operationId: deleteBatch
      tags: [batches]
      summary: Delete batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /batches/{id}/assign:
    post:
      operationId: assignBatchMembers
      tags: [batches]
      summary: Assign students and faculty to batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchAssignInput"
      responses:
        "200":
          description: Assigned

  # ── Lectures ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        deviceFingerprint:
          type: string
    RegisterInput:
      type: object
      required: [email, password, fullName, mobileNumber, role]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        mobileNumber:
          type: string
        role:
          type: string
        profilePhoto:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        interestedCourse:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        mobileNumber:
          type: string
        profilePhoto:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string

    # ── User schemas ───────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, fullName, role, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        mobileNumber:
          type: ["string", "null"]
        role:
          type: string
        status:
          type: string
        profilePhoto:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        interestedCourse:
          type: ["string", "null"]
        experience:
          type: ["string", "null"]
        expertise:
          type: ["string", "null"]
        createdAt:
          type: string
...
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        reason:
          type: ["string", "null"]
    AdminInput:
      type: object
      required: [email, password, fullName, mobileNumber]
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        mobileNumber:
          type: string
    AdminUpdate:
      type: object
      properties:
        fullName:
          type: string
        mobileNumber:
          type: string

    # ── Course schemas ─────────────────────────────────────────────────────
...
    Batch:
      type: object
      required: [id, name, courseId, startDate, endDate, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        courseId:
          type: integer
        courseName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        studentCount:
          type: integer
        facultyCount:
          type: integer
        createdAt:
          type: string
    BatchInput:
      type: object
      required: [name, courseId, startDate, endDate]
      properties:
        name:
          type: string
        courseId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
    BatchUpdate:
      type: object
      properties:
        name:
          type: string
        courseId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
    BatchAssignInput:
      type: object
      properties:
        studentIds:
          type: array
          items:
            type: integer
        facultyIds:
          type: array
          items:
            type: integer

    # ── Lecture schemas ────────────────────────────────────────────────────
...